+++ castle/cms/tests/test_theming.py
@@ -1,15 +1,8 @@
 # -*- coding: utf-8 -*-
-import unittest
-
-from castle.cms import trash
-from castle.cms.interfaces import ITrashed
-from castle.cms.testing import CASTLE_PLONE_INTEGRATION_TESTING
-from plone import api
-from plone.app.testing import TEST_USER_ID
-from plone.app.testing import TEST_USER_NAME
-from plone.app.testing import login
-from plone.app.testing import setRoles
 from castle.cms import theming
+from castle.cms.testing import CASTLE_PLONE_INTEGRATION_TESTING
+
+import unittest
 
 
 MINIMAL_LAYOUT = """<!doctype html>
@@ -72,5 +65,5 @@ class TestTheming(unittest.TestCase):
     def test_apply(self):
         transform = theming.getTransform(self.portal, self.request)
         result = transform(self.request, MINIMAL_RESULT, context=self.portal)
-        self.assertTrue('<div>foobar</div>' in result)
-        self.assertTrue('data-pat-structure' in result)
+        self.assertTrue('<div>foobar</div>' in ''.join(result))
+        self.assertTrue('data-pat-structure' in ''.join(result))