+++ HISTORY.rst
@@ -4,4 +4,6 @@ Changelog
 1.1.0 (unreleased)
 ------------------
 
-- 
+- Fix query filter tile to handle sub-filters that are not in the current
+  base query for the result set
+  [vangheem]